requirements.txt
# Core dependencies
fastapi==0.115.13
uvicorn==0.32.1
pydantic==2.11.0a2
python-dotenv==1.0.1

# WebSocket and real-time data
kiteconnect==6.1.0
websockets==14.2
msgpack==1.0.8

# Authentication and security
PyJWT==2.10.0
python-multipart==0.0.20

# Data processing
pandas==2.2.3
numpy==2.0.2
scikit-learn==1.5.2

# Optional: Redis for caching
redis==5.2.1

# AI and analysis dependencies
absl-py==2.1.0
accelerate==1.7.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.9
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
asttokens==3.0.0
astunparse==1.6.3
attrs==25.1.0
beautifulsoup4==4.13.3
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.2.1
comm==0.2.2
dataclasses-json==0.6.7
debugpy==1.8.12
decorator==5.1.1
einops==0.8.1
executing==2.2.0
fake-useragent==2.0.3
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.6.2
fsspec==2025.5.1
gast==0.6.0
google==3.0.0
google-auth==2.38.0
google-genai==1.3.0
google-pasta==0.2.0
grpcio==1.67.1
h11==0.14.0
h5py==3.13.0
hf-xet==1.1.4
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.0
idna==3.10
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.9.0
langchain==0.3.25
langchain-community==0.3.24
langchain-text-splitters==0.3.8
langsmith==0.3.45
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
multidict==6.4.4
mypy_extensions==1.1.0
namex==0.0.8
narwhals==1.42.1
nest-asyncio==1.6.0
networkx==3.5
npm==0.1.1
opt_einsum==3.4.0
optional-django==0.1.0
optree==0.14.1
orjson==3.10.18
outcome==1.3.0.post0
packaging==24.2
parso==0.8.4
peft==0.15.2
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.6
plotly==6.1.2
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic-settings==2.9.1
pydantic_core==2.29.0
Pygments==2.19.1
PySocks==1.7.1

# Tabular ML
catboost==1.2.5
